Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/animation-frames/callback-exception-expected.txt

CONSOLE MESSAGE: Error: requestAnimationFrameException

PASS requestAnimationFrame callback exceptions are reported to error handler
